Закончить основной сервис, тесты не пройдены, необходим fix и проверка

+++ explore-service/src/main/java/ru/practicum/category/service/CategoryServiceImpl.java
@@ -10,6 +10,7 @@ import ru.practicum.category.dto.NewCategoryDto;
 import ru.practicum.category.dto.mapper.CategoryMapper;
 import ru.practicum.category.model.Category;
 import ru.practicum.category.repository.CategoryRepository;
+import ru.practicum.event.repository.EventRepository;
 import ru.practicum.exception.NotFoundException;
 import ru.practicum.exception.ValidationException;
 
@@ -22,6 +23,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Transactional
     @Override
@@ -41,15 +43,16 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public CategoryDto update(CategoryDto category, Long id) {
         log.info("Update category: {}", category);
-        /** Проверим есть ли такая сущность в бд, после проверим имя на уникальность, иначе выбрасываем кастомные методы исключений*/
+        /** Проверим есть ли такая сущность в бд, после проверим имя на уникальность, иначе выбрасываем исключениe*/
         Category entity = validateExistence(id);
         if (entity != null) {
-            if (!categoryRepository.existsByName(category.getName())) {
+            // если имя категории существует в бд и это не имя нашей категории, выбрасываем исключение.
+            if (categoryRepository.existsByName(category.getName()) && !entity.getName().equals(category.getName())) {
+                throw getValidationException(category.getName());
+            } else {
                 // изменяем поле name у сущности из бд и снова записываем ее, обновляя.
                 entity.setName(category.getName());
                 return CategoryMapper.mapToDto(categoryRepository.save(entity));
-            } else {
-                throw getValidationException(category.getName());
             }
         } else {
             throw getNotFoundException(id);
@@ -62,6 +65,9 @@ public class CategoryServiceImpl implements CategoryService {
         log.info("Delete category: {}", id);
         Category entity = validateExistence(id);
         if (entity != null) {
+           if(eventRepository.existsEventByCategoryId(id)){
+               throw getValidationException("The category is not empty");
+           }
             log.info("Category: {} is deleted", entity);
             categoryRepository.delete(entity);
         } else {
